+++ packages/cli/src/modules/agents/__tests__/agent-validation.service.test.ts
@@ -42,10 +42,9 @@ function makeAiService(proxyEnabled = false) {
 function makeService(aiService = makeAiService()) {
 	const agentRepository = mock<AgentRepository>();
 	const agentSkillsService = mock<AgentSkillsService>();
-	agentSkillsService.getMissingSkillIds.mockReturnValue([]);
 
 	return {
-		service: new AgentValidationService(agentRepository, agentSkillsService, aiService),
+		service: new AgentValidationService(agentRepository, aiService),
 		agentRepository,
 		agentSkillsService,
 		aiService,
@@ -91,8 +90,7 @@ describe('AgentValidationService', () => {
 	});
 
 	it('consolidates missing optional credential-backed features', async () => {
-		const { service, agentRepository, agentSkillsService } = makeService();
-		agentSkillsService.getMissingSkillIds.mockReturnValue(['skill-1']);
+		const { service, agentRepository } = makeService();
 		agentRepository.findByIdAndProjectId.mockResolvedValue(
 			makeAgent({
 				...runnableConfig,
