+++ packages/cli/src/modules/agents/utils/agent-published-snapshot.ts
@@ -13,7 +13,7 @@ export function getPublishedAgentSnapshot(agentEntity: Agent): Agent {
 	return {
 		...agentEntity,
 		schema: activeVersionSchema,
-		tools: agentEntity.activeVersion?.tools ?? agentEntity.tools ?? {},
+		tools: agentEntity.activeVersion?.tools ?? {},
 		skills: agentEntity.activeVersion?.skills ?? {},
 	} as Agent;
 }
